+++ components/layout/layout_task.rs
@@ -48,6 +48,7 @@ use profile_traits::time::{self, TimerMetadata, profile};
 use query::{LayoutRPCImpl, process_content_box_request, process_content_boxes_request};
 use query::{MarginPadding, MarginRetrievingFragmentBorderBoxIterator, PositionProperty};
 use query::{PositionRetrievingFragmentBorderBoxIterator, Side};
+use query::{process_node_geometry_request, process_offset_parent_query, process_resolved_style_request};
 use script::dom::node::LayoutData;
 use script::layout_interface::Animation;
 use script::layout_interface::{LayoutChan, LayoutRPC, OffsetParentResponse};
@@ -817,168 +818,6 @@ impl LayoutTask {
                                               traversal);
     }
 
-    fn process_node_geometry_request(&self,
-                                     requested_node: TrustedNodeAddress,
-                                     layout_root: &mut FlowRef)
-                                     -> Rect<i32> {
-        let requested_node: OpaqueNode = OpaqueNodeMethods::from_script_node(requested_node);
-        let mut iterator = FragmentLocatingFragmentIterator::new(requested_node);
-        sequential::iterate_through_flow_tree_fragment_border_boxes(layout_root, &mut iterator);
-        iterator.client_rect
-    }
-
-    /// Return the resolved value of property for a given (pseudo)element.
-    /// https://drafts.csswg.org/cssom/#resolved-value
-    fn process_resolved_style_request(&self,
-                                      requested_node: TrustedNodeAddress,
-                                      pseudo: &Option<PseudoElement>,
-                                      property: &Atom,
-                                      layout_root: &mut FlowRef)
-                                      -> Option<String> {
-        let node = unsafe { LayoutNode::new(&requested_node) };
-
-        let layout_node = ThreadSafeLayoutNode::new(&node);
-        let layout_node = match pseudo {
-            &Some(PseudoElement::Before) => layout_node.get_before_pseudo(),
-            &Some(PseudoElement::After) => layout_node.get_after_pseudo(),
-            _ => Some(layout_node)
-        };
-
-        let layout_node = match layout_node {
-            None => {
-                // The pseudo doesn't exist, return nothing.  Chrome seems to query
-                // the element itself in this case, Firefox uses the resolved value.
-                // https://www.w3.org/Bugs/Public/show_bug.cgi?id=29006
-                return None;
-            }
-            Some(layout_node) => layout_node
-        };
-
-        let style = &*layout_node.style();
-
-        let positioned = match style.get_box().position {
-            position::computed_value::T::relative |
-            /*position::computed_value::T::sticky |*/
-            position::computed_value::T::fixed |
-            position::computed_value::T::absolute => true,
-            _ => false
-        };
-
-        //TODO: determine whether requested property applies to the element.
-        //      eg. width does not apply to non-replaced inline elements.
-        // Existing browsers disagree about when left/top/right/bottom apply
-        // (Chrome seems to think they never apply and always returns resolved values).
-        // There are probably other quirks.
-        let applies = true;
-
-        fn used_value_for_position_property(layout_node: ThreadSafeLayoutNode,
-                                            layout_root: &mut FlowRef,
-                                            requested_node: TrustedNodeAddress,
-                                            property: &Atom) -> Option<String> {
-            let layout_data = layout_node.borrow_layout_data();
-            let position = layout_data.as_ref().map(|layout_data| {
-                match layout_data.data.flow_construction_result {
-                    ConstructionResult::Flow(ref flow_ref, _) =>
-                        flow::base(flow_ref.deref()).stacking_relative_position,
-                    // TODO(dzbarsky) search parents until we find node with a flow ref.
-                    // https://github.com/servo/servo/issues/8307
-                    _ => ZERO_POINT
-                }
-            }).unwrap_or(ZERO_POINT);
-            let property = match *property {
-                atom!("bottom") => PositionProperty::Bottom,
-                atom!("top") => PositionProperty::Top,
-                atom!("left") => PositionProperty::Left,
-                atom!("right") => PositionProperty::Right,
-                atom!("width") => PositionProperty::Width,
-                atom!("height") => PositionProperty::Height,
-                _ => unreachable!()
-            };
-            let requested_node: OpaqueNode =
-                OpaqueNodeMethods::from_script_node(requested_node);
-            let mut iterator =
-                PositionRetrievingFragmentBorderBoxIterator::new(requested_node,
-                                                                 property,
-                                                                 position);
-            sequential::iterate_through_flow_tree_fragment_border_boxes(layout_root,
-                                                                        &mut iterator);
-            iterator.result.map(|r| r.to_css_string())
-        }
-
-        // TODO: we will return neither the computed nor used value for margin and padding.
-        // Firefox returns blank strings for the computed value of shorthands,
-        // so this should be web-compatible.
-        match *property {
-            atom!("margin-bottom") | atom!("margin-top") |
-            atom!("margin-left") | atom!("margin-right") |
-            atom!("padding-bottom") | atom!("padding-top") |
-            atom!("padding-left") | atom!("padding-right")
-            if applies && style.get_box().display != display::computed_value::T::none => {
-                let (margin_padding, side) = match *property {
-                    atom!("margin-bottom") => (MarginPadding::Margin, Side::Bottom),
-                    atom!("margin-top") => (MarginPadding::Margin, Side::Top),
-                    atom!("margin-left") => (MarginPadding::Margin, Side::Left),
-                    atom!("margin-right") => (MarginPadding::Margin, Side::Right),
-                    atom!("padding-bottom") => (MarginPadding::Padding, Side::Bottom),
-                    atom!("padding-top") => (MarginPadding::Padding, Side::Top),
-                    atom!("padding-left") => (MarginPadding::Padding, Side::Left),
-                    atom!("padding-right") => (MarginPadding::Padding, Side::Right),
-                    _ => unreachable!()
-                };
-                let requested_node: OpaqueNode =
-                    OpaqueNodeMethods::from_script_node(requested_node);
-                let mut iterator =
-                    MarginRetrievingFragmentBorderBoxIterator::new(requested_node,
-                                                                   side,
-                                                                   margin_padding,
-                                                                   style.writing_mode);
-                sequential::iterate_through_flow_tree_fragment_border_boxes(layout_root,
-                                                                            &mut iterator);
-                iterator.result.map(|r| r.to_css_string())
-            },
-
-            atom!("bottom") | atom!("top") | atom!("right") |
-            atom!("left")
-            if applies && positioned && style.get_box().display !=
-                    display::computed_value::T::none => {
-                used_value_for_position_property(layout_node, layout_root, requested_node, property)
-            }
-            atom!("width") | atom!("height")
-            if applies && style.get_box().display !=
-                    display::computed_value::T::none => {
-                used_value_for_position_property(layout_node, layout_root, requested_node, property)
-            }
-            // FIXME: implement used value computation for line-height
-            ref property => {
-                style.computed_value_to_string(property.as_slice()).ok()
-            }
-        }
-    }
-
-    fn process_offset_parent_query(&self,
-                                   requested_node: TrustedNodeAddress,
-                                   layout_root: &mut FlowRef)
-                                   -> OffsetParentResponse {
-        let requested_node: OpaqueNode = OpaqueNodeMethods::from_script_node(requested_node);
-        let mut iterator = ParentOffsetBorderBoxIterator::new(requested_node);
-        sequential::iterate_through_flow_tree_fragment_border_boxes(layout_root, &mut iterator);
-        let parent_info_index = iterator.parent_nodes.iter().rposition(|info| info.is_some());
-        match parent_info_index {
-            Some(parent_info_index) => {
-                let parent = iterator.parent_nodes[parent_info_index].as_ref().unwrap();
-                let origin = iterator.node_border_box.origin - parent.border_box.origin;
-                let size = iterator.node_border_box.size;
-                OffsetParentResponse {
-                    node_address: Some(parent.node_address.to_untrusted_node_address()),
-                    rect: Rect::new(origin, size),
-                }
-            }
-            None => {
-                OffsetParentResponse::empty()
-            }
-        }
-    }
-
     fn compute_abs_pos_and_build_display_list<'a>(&'a mut self,
                                                   data: &Reflow,
                                                   layout_root: &mut FlowRef,
@@ -1220,16 +1059,12 @@ impl LayoutTask {
                 ReflowQueryType::ContentBoxesQuery(node) =>
                     rw_data.content_boxes_response = process_content_boxes_request(node, &mut root_flow),
                 ReflowQueryType::NodeGeometryQuery(node) =>
-                    rw_data.client_rect_response = self.process_node_geometry_request(node, &mut root_flow),
-                ReflowQueryType::ResolvedStyleQuery(node, ref pseudo, ref property) => {
+                    rw_data.client_rect_response = process_node_geometry_request(node, &mut root_flow),
+                ReflowQueryType::ResolvedStyleQuery(node, ref pseudo, ref property) =>
                     rw_data.resolved_style_response =
-                        self.process_resolved_style_request(node,
-                                                            pseudo,
-                                                            property,
-                                                            &mut root_flow)
-                }
+                    process_resolved_style_request(node, pseudo, property, &mut root_flow),
                 ReflowQueryType::OffsetParentQuery(node) =>
-                    rw_data.offset_parent_response = self.process_offset_parent_query(node, &mut root_flow),
+                    rw_data.offset_parent_response = process_offset_parent_query(node, &mut root_flow),
                 ReflowQueryType::NoQuery => {}
             }
         }
@@ -1474,7 +1309,7 @@ impl LayoutTask {
     }
 }
 
-struct FragmentLocatingFragmentIterator {
+pub struct FragmentLocatingFragmentIterator {
     node_address: OpaqueNode,
     client_rect: Rect<i32>,
 }
